HackingScripts/Python-Scripts/requirements.txt
beautifulsoup4==4.10.0
dnslib==0.9.16
dulwich==0.20.26
exif==1.3.4
impacket==0.9.23
name_that_hash==1.10.0
netifaces==0.10.9
paramiko==2.7.1
Pillow==9.0.0
pwn==1.0
pwntools==4.7.0
requests==2.23.0
SocksiPy_branch==1.01
urllib3==1.25.11
argparse
mmh3
requests
shodan
tldextract
termcolor
netaddr
colored
goop
lockfile
pythonwhois
multiprocessing
queuelib










